ajustando para reportar erro

+++ src/condutores/condutores.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreateCondutoreDto } from './dto/create-condutore.dto';
 import { UpdateCondutoreDto } from './dto/update-condutore.dto';
 import { PrismaService } from 'prisma/prisma.service';
@@ -34,10 +34,24 @@ export class CondutoresService {
   }
 
   async remove(id: number) {
+    const pesquisa = await this.prisma.solicitacao_condutores.findMany({
+      where: {
+        condutores_id: id,
+      },
+    });
+  
+    if (pesquisa.length > 0) {
+      throw new HttpException(
+        'Não é possível excluir este condutor, pois ele está vinculado a uma solicitação de viagem.',
+        HttpStatus.CONFLICT, // 409 Conflict
+      );
+    }
+  
     return await this.prisma.condutores.delete({
       where: {
-        id: id
-      }
-    })
+        id: id,
+      },
+    });
   }
+  
 }
